+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Documentation improvements
 * CI improvements
 * `update_particle_indexer_new_lower_count!`, `delete_particle_end_group2!` now set `end2` to -1 if last particle from group2 is deleted
+* Now possible to write particle data via `write_netcdf`
 
 ## v0.7.0
 * Added use of MuladdMacro (via `@muladd`), updated tolerances in tests and reference solutions